# server/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CampAI",
        "version": "1.0.0"
    },
    "paths":{
...
        "/users":{
            "post":{
                "description": "Creates a user",
                "parameters": [{
                    "in": "path",
                    "name":"body",
                    "description": "User object",
                    "required": true,
                    "schema": {
                        "$ref": "#/components/schemas/User"
                    }
                }],
                "responses": {
                    "200":{
                        "description": "Successfuly created user#ID"
                    },
                    "500":{
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete":{
                "description": "Delete a single dog",
                "responses": {
                    "200":{
                        "description": "Dog successfully deleted"
                    },
                    "400":{
                        "description": "The API received a bad parameter."
                    }
                },
                "parameters": [{
                    "in": "path",
                    "name":"body",
                    "description": "Dog object",
                    "required": true,
                    "schema": {
                        "$ref": "#/components/schemas/User"
                    }
                }]
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}